==> scripts/7.generate_structural_score.py <==
''' Import libraries '''
import pandas as pd
import numpy as np
np.random.seed(8)

import _pickle as cPickle
import os
import glob
import time
import argparse

#parser = argparse.ArgumentParser(description='### The script for predicting probabilities of proteolytic events in proteins by structure information (structures generated from PDB or AlphaFoldDB databases) ###')
#parser.add_argument("-s", "--structure", help="The input structure will be processed by structural model for experimental structures from PDB database", action="store_true")
#parser.add_argument("-m", "--model", help="The input structure will be processed by structural model for AlphaFold structures from AlphaFoldDB database", action="store_true")
#parser.add_argument("-b", "--batch", help="The input structure will be processed by both structural models", action="store_true")
#args = parser.parse_args()

''' Set paths ''' 
main_path = os.getcwd()
model_path = os.path.join(main_path, "models")
StructureSet_path = os.path.join(main_path, "results")
structure_list = [i for i in os.listdir(StructureSet_path) if '.' not in i]

''' Features '''
features_exp = {"ACC_N_chain":float, "bfac_N_chain":float, "SS_B":int, "SS_H":int, "SS_O":int, "len_loop_N_chain":float, "is_terminus":int}
features_AF = {"ACC_N_chain":float, "confidence_score_N_chain":float, "SS_B":int, "SS_H":int, "SS_O":int, "len_loop_N_chain":float, "is_terminus":int}

''' Loading  Models '''
# Structural model #
with open(os.path.join(model_path, "structural_model.pkl"), 'rb') as file:
    structural_model = cPickle.load(file)
# Structural model for AlphaFold structures #
with open(os.path.join(model_path, "structural_model_AlphaFold.pkl"), 'rb') as file:
    structural_model_AF = cPickle.load(file)

def main():
    
    num = 1
    for structure in structure_list:        
        structure_path = os.path.join(StructureSet_path, structure)
        feature_path = os.path.join(structure_path, "features")
        
        feature_files = glob.glob(os.path.join(feature_path, '*.norm.csv'))
        for feature_file in feature_files:
            structure_name = feature_file.split('/')[-1].split('.norm.csv')[0]
            
            ''' Get structural score '''
            if 'AF-' in structure_name:
                feature_df = pd.read_csv(feature_file, dtype=features_AF)
                X_test = feature_df[features_AF.keys()].values
                structural_scores = structural_model_AF.predict_proba(X_test)[:, 1]
                print(f"{num} --- {structure_name} --- Predictions generated by AF structural model")
            else:
                feature_df = pd.read_csv(feature_file, dtype=features_exp)
                X_test = feature_df[features_exp.keys()].values
                structural_scores = structural_model.predict_proba(X_test)[:, 1]
                print(f"{num} --- {structure_name} --- Predictions generated by structural model")

            score_path = os.path.join(structure_path, "scores")
            if not os.path.exists(score_path):
                os.mkdir(score_path)
            
            structural_score_col = f"PredictedProbability"
            feature_df[structural_score_col] = structural_scores
            feature_df[structural_score_col] = feature_df[structural_score_col].map(lambda x: round(x, 4))
            
            feature_df.to_csv(os.path.join(score_path, f"{structure_name}.StrProba.csv"), index=False)
            feature_df[["structure", "chain", "AA", "num_AA", "PredictedProbability"]].to_csv(os.path.join(structure_path, f"{structure_name}.predictions.csv"), index=False)
            
            num += 1
            
''' Launch script '''                
if __name__ == "__main__":
    start = time.ctime()
    main()
    finish = time.ctime()
    #print(f"\nScript: {__file__}\nStart: {start}\nFinish: {finish}")
